+++ lib/block_layer.c
@@ -33,7 +33,7 @@ bool init_superblock(){
 bool init_inode_list(){
     // Indirect blocks will use Data blocks itself to store the addresses.
     struct iNode inode[sizeof(struct iNode)];
-    for(int i=0; i<DIRECT_B_COUNT; i++){
+    for(ssize_t i=0; i<DIRECT_B_COUNT; i++){
         inode->direct_blocks[i] = 0; // using 0 because 0 can only be used by superblock
     }
     inode->single_indirect = 0;
@@ -44,17 +44,17 @@ bool init_inode_list(){
     inode->num_blocks = 0;
     inode->allocated = false;
     char buff[BLOCK_SIZE];
-    int dummy_value = 0;
-    for(int block_id=1; block_id<=INODE_B_COUNT; block_id++){
-        int offset = 0;
+    ssize_t dummy_value = 0;
+    for(ssize_t block_id=1; block_id<=INODE_B_COUNT; block_id++){
+        ssize_t offset = 0;
         memset(buff, dummy_value, BLOCK_SIZE);
         // each i-node block will be initialized with stuct inode. 
-        for(int i=0; i<super_block->inodes_per_block; i++){
+        for(ssize_t i=0; i<super_block->inodes_per_block; i++){
             memcpy(buff+offset, inode, sizeof(struct iNode));
             offset += sizeof(struct iNode);
         }
         if(!write_block(block_id, buff)){    
-            printf("Could not create inode at block - %d", block_id);
+            printf("Could not create inode at block - %ld", block_id);
             return false;
         }
     }
@@ -63,17 +63,17 @@ bool init_inode_list(){
 
 bool init_freelist(){
     char buff[BLOCK_SIZE];
-    int block_id = super_block->free_list_head;
-    int dummy_value = 0;
+    ssize_t block_id = super_block->free_list_head;
+    ssize_t dummy_value = 0;
      // 0th index will store the address of the next free list block that holds the addresses. 
     // Eg: Block 100:
     // 101-612, Block[0]=613
-    for(int i=0; i<FREE_LIST_BLOCKS; i++){
-        int curr_block_id = block_id;
-        // printf("Started with %d and ", curr_block_id);
+    for(ssize_t i=0; i<FREE_LIST_BLOCKS; i++){
+        ssize_t curr_block_id = block_id;
+        // printf("Started with %ld and ", curr_block_id);
         memset(buff, dummy_value, BLOCK_SIZE);
-        int offset = 0;
-        for(int j=1; j<DBLOCKS_PER_BLOCK; j++){
+        ssize_t offset = 0;
+        for(ssize_t j=1; j<DBLOCKS_PER_BLOCK; j++){
             offset += ADDRESS_SIZE;
             block_id++;
             if(block_id >= BLOCK_COUNT){
@@ -89,7 +89,7 @@ bool init_freelist(){
         } else{
             memcpy(buff, &block_id, ADDRESS_SIZE);
         }
-        // printf("ended with %d\n", block_id);
+        // printf("ended with %ld\n", block_id);
         if(!write_block(curr_block_id, buff)){
             printf("Unable to write in the free list");
             return false;
@@ -101,7 +101,7 @@ bool init_freelist(){
     return true;
 }
 
-int create_new_dblock(){
+ssize_t create_new_dblock(){
     if(super_block->free_list_head == 0)
     {
         printf("No more blocks left\n");
@@ -111,28 +111,28 @@ int create_new_dblock(){
     if(!read_block(super_block->free_list_head, buff)){
         return -1;
     }
-    int *dblock_num_ptr = (int *)buff;
-    int ans = 0;
+    ssize_t *dblock_num_ptr = (ssize_t *)buff;
+    ssize_t ans = 0;
     //DBLOCKS_PER_BLOCK gives the number of block addresses a block can hold.
     // iterating from 1st index because 0th index is used to point to the next free node.
-    for(int i=1; i<DBLOCKS_PER_BLOCK; i++){
-        int dblock_num = dblock_num_ptr[i];
+    for(ssize_t i=1; i<DBLOCKS_PER_BLOCK; i++){
+        ssize_t dblock_num = dblock_num_ptr[i];
         if(dblock_num_ptr[i]!=0){
-            // printf("pos - %d : %d\n", i, dblock_num);
+            // printf("pos - %ld : %ld\n", i, dblock_num);
             // the dblock is free and we are allotting this
             dblock_num_ptr[i]=0;
             ans = dblock_num;
             break;
         }
     }
-    int temp = super_block->free_list_head;
+    ssize_t temp = super_block->free_list_head;
     if(ans == 0){
         //freeList node itself is allocated as DataBlock
         ans = super_block->free_list_head;
         // move free_list_head to the next free_list_block
         super_block->free_list_head = dblock_num_ptr[0];
         dblock_num_ptr[0] = 0;
-        printf("Freelist Node %d is now a dblock\n", ans);
+        printf("Freelist Node %ld is now a dblock\n", ans);
         if(!write_superblock()){
             return -1;
         }
@@ -149,9 +149,9 @@ int create_new_dblock(){
     return ans;
 }
 
-char *read_dblock(int dblock_num){
+char *read_dblock(ssize_t dblock_num){
     if(dblock_num <= INODE_B_COUNT || dblock_num > BLOCK_COUNT){
-        printf("Invalid data block number %d provided", dblock_num);
+        printf("Invalid data block number %ld provided", dblock_num);
         return NULL;
     }
     // reading the block data to the buffer and returning it.
@@ -162,18 +162,18 @@ char *read_dblock(int dblock_num){
     return NULL;
 }
 
-bool write_dblock(int dblock_num, char *buff){
+bool write_dblock(ssize_t dblock_num, char *buff){
     if(dblock_num <= INODE_B_COUNT || dblock_num > BLOCK_COUNT){
-        printf("Invalid data block number %d provided\n", dblock_num);
+        printf("Invalid data block number %ld provided\n", dblock_num);
         return false;
     }
     //writing the block of data through the disk layer write_block call.
     return write_block(dblock_num, buff);
 }
 
-bool free_dblock(int dblock_num){
+bool free_dblock(ssize_t dblock_num){
     if(dblock_num <= INODE_B_COUNT || dblock_num > BLOCK_COUNT){
-        printf("Invalid data block number %d provided\n", dblock_num);
+        printf("Invalid data block number %ld provided\n", dblock_num);
         return false;
     }
     char buff[BLOCK_SIZE];
@@ -190,9 +190,9 @@ bool free_dblock(int dblock_num){
     if(!read_block(super_block->free_list_head, buff)){
         return false;
     }
-    int* dblock_num_ptr = (int*) buff;
-    int ans = 0;
-    for(int i=1; i<DBLOCKS_PER_BLOCK; i++){
+    ssize_t* dblock_num_ptr = (ssize_t*) buff;
+    ssize_t ans = 0;
+    for(ssize_t i=1; i<DBLOCKS_PER_BLOCK; i++){
         // if the free list head have one of the entry as 0, then that place can be updated with the block address of newly released block
         if(dblock_num_ptr[i]==0){
             dblock_num_ptr[i] = dblock_num;
@@ -204,9 +204,9 @@ bool free_dblock(int dblock_num){
     // the newly released dblock becomes the head of free list.
     if(ans==0){
         // dblock will be made the new free list which points to old one
-        int temp = super_block->free_list_head;
+        ssize_t temp = super_block->free_list_head;
         super_block->free_list_head = dblock_num;
-        printf("Dblock %d is named as the free list head\n", dblock_num);
+        printf("Dblock %ld is named as the free list head\n", dblock_num);
         if(!write_superblock()){
             return false;
         }
@@ -226,33 +226,33 @@ bool free_dblock(int dblock_num){
 }
 
 //helper functions
-bool is_valid_inum(int inode_num){
+bool is_valid_inum(ssize_t inode_num){
     return inode_num>0 && inode_num<super_block->inode_count;
 }
 
 //which block contains the corresponding inode_num
-int inode_num_to_block_id(int inode_num){
+ssize_t inode_num_to_block_id(ssize_t inode_num){
     return 1 + (inode_num/super_block->inodes_per_block);
 }
 
 // offset with the blockId that points to the struct of inode_num
-int inode_num_to_offset(int inode_num){
+ssize_t inode_num_to_offset(ssize_t inode_num){
     return inode_num % super_block->inodes_per_block;
 }
 
-int create_new_inode(){
+ssize_t create_new_inode(){
     // TODO: Worst case - O(N) for finding one even with use of latest inum
     if(!is_valid_inum(super_block->latest_inum)){
-        printf("Invalid inode num - %d being accessed or out of range during create_new_inode\n", super_block->latest_inum);
+        printf("Invalid inode num - %ld being accessed or out of range during create_new_inode\n", super_block->latest_inum);
         return -1;
     }
     // latest-inum is the next inum value from the prev allocated inum. Being used to improve the search.
-    int block_id = inode_num_to_block_id(super_block->latest_inum);
-    int offset = inode_num_to_offset(super_block->latest_inum);
+    ssize_t block_id = inode_num_to_block_id(super_block->latest_inum);
+    ssize_t offset = inode_num_to_offset(super_block->latest_inum);
     char buff[BLOCK_SIZE];
     struct iNode* temp = NULL;
     struct iNode* inode = NULL;
-    int visit_count = 0;
+    ssize_t visit_count = 0;
     while(visit_count != super_block->inode_count){
         // reading block
         if(!read_block(block_id, buff)){
@@ -264,7 +264,7 @@ int create_new_inode(){
             visit_count++;
             if(inode->allocated==false){
                 //allocating the corresponding inode.
-                int ans = super_block->latest_inum;
+                ssize_t ans = super_block->latest_inum;
                 // getting the new latest inum from block_id and offset
                 super_block->latest_inum += visit_count;
                 inode->allocated = true;
@@ -280,15 +280,15 @@ int create_new_inode(){
     return -1;
 }
 
-struct iNode* read_inode(int inode_num){
+struct iNode* read_inode(ssize_t inode_num){
     if(!is_valid_inum(inode_num)){
-        printf("Invalid inode num - %d being accessed or out of range\n", inode_num);
+        printf("Invalid inode num - %ld being accessed or out of range\n", inode_num);
         return false;
     }
     //get BlockId which contains the inode.
-    int block_id = inode_num_to_block_id(inode_num);
+    ssize_t block_id = inode_num_to_block_id(inode_num);
     // get offset within the blockId to point to exact inode struct
-    int offset = inode_num_to_offset(inode_num);
+    ssize_t offset = inode_num_to_offset(inode_num);
     char buff[BLOCK_SIZE];
     if(!read_block(block_id, buff)){
         return false;
@@ -301,15 +301,15 @@ struct iNode* read_inode(int inode_num){
     return ans;
 }
 
-bool write_inode(int inode_num, struct iNode* inode){
+bool write_inode(ssize_t inode_num, struct iNode* inode){
     if(!is_valid_inum(inode_num)){
-        printf("Invalid inode num - %d being accessed or out of range\n", inode_num);
+        printf("Invalid inode num - %ld being accessed or out of range\n", inode_num);
         return false;
     }
     //get BlockId which contains the inode.
-    int block_id = inode_num_to_block_id(inode_num);
+    ssize_t block_id = inode_num_to_block_id(inode_num);
     // get offset within the blockId to point to exact inode struct
-    int offset = inode_num_to_offset(inode_num);
+    ssize_t offset = inode_num_to_offset(inode_num);
     char buff[BLOCK_SIZE];
     if(!read_block(block_id, buff)){
         return false;
@@ -321,7 +321,7 @@ bool write_inode(int inode_num, struct iNode* inode){
     if(!write_block(block_id, buff)){
         return false;
     }
-    DEBUG_PRINTF("Wrote inode %d to block %d offset %d", inode_num, block_id, offset);
+    DEBUG_PRINTF("Wrote inode %ld to block %ld offset %ld", inode_num, block_id, offset);
     return true;
 }
 
@@ -330,11 +330,11 @@ bool free_dblocks_from_inode(struct iNode* inode){
     char dblock_list_buff[BLOCK_SIZE];
     char single_indirect_block_buff[BLOCK_SIZE];
     char double_indirect_block_buff[BLOCK_SIZE];
-    int* dblock_list_ptr;
-    int* single_indirect_block_ptr;
-    int* double_indirect_block_ptr;
+    ssize_t* dblock_list_ptr;
+    ssize_t* single_indirect_block_ptr;
+    ssize_t* double_indirect_block_ptr;
     // freeing direct block
-    for(int i=0; i<DIRECT_B_COUNT; i++){
+    for(ssize_t i=0; i<DIRECT_B_COUNT; i++){
         if(inode->direct_blocks[i]==0){
             done = true;
             break;
@@ -351,8 +351,8 @@ bool free_dblocks_from_inode(struct iNode* inode){
         return false;
     }
     // store dblock number each by int
-    dblock_list_ptr = (int* )dblock_list_buff;
-    for(int i=0; i<DBLOCKS_PER_BLOCK; i++){
+    dblock_list_ptr = (ssize_t* )dblock_list_buff;
+    for(ssize_t i=0; i<DBLOCKS_PER_BLOCK; i++){
         if(dblock_list_ptr[i]==0){
             done = true;
             break;
@@ -372,8 +372,8 @@ bool free_dblocks_from_inode(struct iNode* inode){
     if(!read_block(inode->double_indirect, single_indirect_block_buff)){
         return false;
     }
-    single_indirect_block_ptr = (int* )single_indirect_block_buff;
-    for(int i=0; i<DBLOCKS_PER_BLOCK; i++){
+    single_indirect_block_ptr = (ssize_t* )single_indirect_block_buff;
+    for(ssize_t i=0; i<DBLOCKS_PER_BLOCK; i++){
         if(single_indirect_block_ptr[i]==0){
             done = true;
             break;
@@ -382,8 +382,8 @@ bool free_dblocks_from_inode(struct iNode* inode){
         if(!read_block(single_indirect_block_ptr[i], dblock_list_buff)){
             return false;
         }
-        dblock_list_ptr = (int* )dblock_list_buff;
-        for(int j=0; j<DBLOCKS_PER_BLOCK; j++){
+        dblock_list_ptr = (ssize_t* )dblock_list_buff;
+        for(ssize_t j=0; j<DBLOCKS_PER_BLOCK; j++){
             if(dblock_list_ptr[j]==0){
                 done = true;
                 break;
@@ -406,8 +406,8 @@ bool free_dblocks_from_inode(struct iNode* inode){
     if(!read_block(inode->triple_indirect, double_indirect_block_buff)){
         return false;
     }
-    double_indirect_block_ptr = (int* )double_indirect_block_buff;
-    for(int k=0; k<DBLOCKS_PER_BLOCK; k++){
+    double_indirect_block_ptr = (ssize_t* )double_indirect_block_buff;
+    for(ssize_t k=0; k<DBLOCKS_PER_BLOCK; k++){
         if(double_indirect_block_ptr[k]==0){
             done = true;
             break;
@@ -415,8 +415,8 @@ bool free_dblocks_from_inode(struct iNode* inode){
         if(!read_block(double_indirect_block_ptr[k], single_indirect_block_buff)){
             return false;
         }
-        single_indirect_block_ptr = (int* )single_indirect_block_buff;
-        for(int i=0; i<DBLOCKS_PER_BLOCK; i++){
+        single_indirect_block_ptr = (ssize_t* )single_indirect_block_buff;
+        for(ssize_t i=0; i<DBLOCKS_PER_BLOCK; i++){
             if(single_indirect_block_ptr[i]==0){
                 done = true;
                 break;
@@ -424,8 +424,8 @@ bool free_dblocks_from_inode(struct iNode* inode){
             if(!read_block(single_indirect_block_ptr[i], dblock_list_buff)){
                 return false;
             }
-            dblock_list_ptr = (int* )dblock_list_buff;
-            for(int j=0; j<DBLOCKS_PER_BLOCK; j++){
+            dblock_list_ptr = (ssize_t* )dblock_list_buff;
+            for(ssize_t j=0; j<DBLOCKS_PER_BLOCK; j++){
                 if(dblock_list_ptr[j]==0){
                     done = true;
                     break;
@@ -445,13 +445,13 @@ bool free_dblocks_from_inode(struct iNode* inode){
     return true;
 }
 
-bool free_inode(int inode_num){
+bool free_inode(ssize_t inode_num){
     if(!is_valid_inum(inode_num)){
-        printf("Invalid inode num - %d being accessed or out of range\n", inode_num);
+        printf("Invalid inode num - %ld being accessed or out of range\n", inode_num);
         return false;
     }
-    int block_id = inode_num_to_block_id(inode_num);
-    int offset = inode_num_to_offset(inode_num);
+    ssize_t block_id = inode_num_to_block_id(inode_num);
+    ssize_t offset = inode_num_to_offset(inode_num);
     char buff[BLOCK_SIZE];
     if(!read_block(block_id, buff)){
         return false;
@@ -467,7 +467,7 @@ bool free_inode(int inode_num){
     if(!write_block(block_id, buff)){
         return false;
     }
-    DEBUG_PRINTF("Inode %d has been freed \n", inode_num);
+    DEBUG_PRINTF("Inode %ld has been freed \n", inode_num);
     return true;
 }
